requirements.txt
docopt
docutils
pyaml
pytest
colorlog